Create ML project
This HTML document is a simple web page that demonstrates image classification using the MobileNet model with the ml5.js library. Here's a breakdown of the code:

<!DOCTYPE html>: Declares the document type and version of HTML being used.

<html>: The root element of the HTML document.

<script> tags: These are used to include external JavaScript libraries: p5.js and ml5.js. p5.js is a library for creative coding and ml5.js is a machine learning library built on top of TensorFlow.js.

<body>: Contains the content of the web page.

<h1> and <h2> tags: These are heading elements used to display titles on the page.

<div id="result">: This is a division element with an ID of "result" where the classification result will be displayed.

<img id="image" src="pic1.jpg" width="100%">: This is an image element with an ID of "image". It displays an image with the filename "pic1.jpg" and takes up 100% of the width of its container.

Inside the <script> tag:

const classifier = ml5.imageClassifier('MobileNet');: This line initializes an image classifier using the MobileNet model provided by ml5.js.
classifier.classify(document.getElementById("image"), gotResult);: This line classifies the image with the ID "image" using the initialized classifier. When the classification result is obtained, it calls the gotResult function.
function gotResult(error, results) { ... }: This is the gotResult function that receives the classification result. If there's an error, it displays the error message in the "result" div. Otherwise, it displays the label and confidence score of the top result in the "result" div.
Overall, this code sets up a web page where you can upload an image (pic1.jpg in this case) and it will classify the image using the MobileNet model, displaying the result and confidence score on the page.